backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "pytz>=2024.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
